TextureItem.CopyTexture to TextureConverter.CopyTexture
