--- Assets/UniGLTF/Runtime/UniGLTF/MaterialFacotry/TextureItem.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 09cdeb503b1e8234faf528b6ae134e7a
-timeCreated: 1519799583
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
